refactor: move toScoredForecastContext mapper from adapters to app layer
Moves the toScoredForecastContext mapper from src/adapters/n8n/mappers/ to
src/app/run-photo-brief/mappers/ to eliminate the unwanted dependency from
app layer to adapter layer.

Changes:
- Moved to-scored-forecast.ts to app layer and updated import paths
- Updated finalize-brief.ts to import from new local location
- Updated render-outputs.ts (adapter) to import from app layer

The remaining import from adapters in the mapper is a type-only import for
RenderableRuntimeContext, which defines the input boundary from n8n. This is
acceptable as it's a type definition, not a runtime dependency.

Fixes #125
